Fixed: Coop player setup (mostly ClientID detection) and update notice logic
Fixes player setup timing checks, ensures update notices are only sent to outdated clients
Improves debug logging for player version and data restoration.
Also corrects the type for coopSetupNextCheckTime from int to float and removes redundant or commented-out code.

+++ coop/code/coop_manager.hpp
@@ -227,7 +227,7 @@ public:
     int getPlayerData_coopSetupTries(Player* player);
     void setPlayerData_coopSetupTries_update(Player* player);
     void setPlayerData_coopSetupTries_reset(Player* player);
-    int getPlayerData_coopSetupNextCheckTime(Player* player);
+    float getPlayerData_coopSetupNextCheckTime(Player* player);
     void setPlayerData_coopSetupNextCheckTime_update(Player* player);
     void setPlayerData_coopSetupNextCheckTime_reset(Player* player);
     int getPlayerData_coopAdminAuthAttemps(Player* player);
